Clean up registration component
